don't split api and remote useCases anymore
