+++ app/src/main/java/com/imdbmovieapp/domain/usecase/GetFavoriteMoviesUseCaseImpl.kt
@@ -1,4 +1,4 @@
-package com.imdbmovieapp.domain.usecase.local
+package com.imdbmovieapp.domain.usecase
 
 import com.imdbmovieapp.domain.model.FavoriteMovieDomain
 import com.imdbmovieapp.domain.repository.FavoriteMovieRepository
